requirements.txt
google-genai>=1.56.0
python-dotenv>=1.0.0

